+++ CHANGELOG.md
@@ -6,8 +6,11 @@ All notable changes to this project are documented in this file.
 ### Added
 
 - Oracle databases are now supported via driver `oracledb`. [#5]
+- DuckDB databases are now supported via driver `duckdb`. [#14]
 - Added support for `decimal.Decimal` type. [#13]
-- (INTERNAL) `BaseDriver` subclasses can now optionally implement an `init_connection` method which is called right after a `PyDbAPIv2Connection` is created so as to perform any driver specific initializations on said connection. [#5]
+- (INTERNAL) `BaseDriver` subclasses can now optionally implement an `init_connection` method which is called right after a `PyDbAPIv2Connection` is created so as to perform any driver-specific initialization steps on said connection. [#5,#14]
+- (INTERNAL) `BaseDriver` subclasses can now optionally implement an `init_transaction` method which is called right before a
+transaction is to be started so as to perform any driver-specific initialization steps regarding the transaction. [#14]
 
 
 ### Changed
@@ -38,3 +41,4 @@ separate method `__cursor`. [#10]
 [#10]: https://github.com/manoss96/onlymaps/pull/10
 [#11]: https://github.com/manoss96/onlymaps/pull/11
 [#13]: https://github.com/manoss96/onlymaps/pull/13
+[#14]: https://github.com/manoss96/onlymaps/pull/14